Refactored with new components
