+++ package.json
@@ -23,6 +23,7 @@
     "react": "^18.2.0",
     "react-device-detect": "^2.2.3",
     "react-dom": "^18.2.0",
+    "react-hotkeys-hook": "^4.4.1",
     "react-icons": "^4.12.0",
     "react-use": "^17.4.2",
     "tailwind-merge": "^2.1.0"
